// supabase/_shared/lib/types.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export type Database = {
  public: {
    Tables: {
// ...
      messages: {
        Row: {
          assistant_id: string | null;
          completed: boolean;
          content: string | null;
          created_at: string;
          id: string;
          metadata: Json | null;
          role: string;
          thread_id: string;
          user_id: string | null;
        };
        Insert: {
          assistant_id?: string | null;
          completed?: boolean;
          content?: string | null;
          created_at?: string;
          id?: string;
          metadata?: Json | null;
          role: string;
          thread_id: string;
          user_id?: string | null;
        };
        Update: {
          assistant_id?: string | null;
          completed?: boolean;
          content?: string | null;
          created_at?: string;
          id?: string;
          metadata?: Json | null;
          role?: string;
          thread_id?: string;
          user_id?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "messages_assistant_id_fkey";
            columns: ["assistant_id"];
            isOneToOne: false;
            referencedRelation: "assistants";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "messages_thread_id_fkey";
            columns: ["thread_id"];
            isOneToOne: false;
            referencedRelation: "threads";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      email_priority: "LOW" | "MID" | "HIGH";
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};